# swagger.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Job Application Tracker API",
        "version": "1.0.0",
        "description": "REST API for managing companies, job applications, contacts, and interviews."
    },
    "servers": [
        {
            "url": "/",
            "description": "Current server"
        }
    ],
...
    "paths": {
...
        "/contacts": {
            "get": {
                "tags": [
                    "Contacts"
                ],
                "summary": "Get all contacts",
                "operationId": "getAllContacts",
                "responses": {
                    "200": {
                        "description": "Contacts retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Contact"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                }
            },
            "post": {
                "tags": [
                    "Contacts"
                ],
                "summary": "Create a contact",
                "operationId": "createContact",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ContactInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Contact created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Contact"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/BadRequest"
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    }
                },
                "security": [
                    {
                        "sessionCookie": []
                    }
                ]
            }
        },
        "/contacts/{id}": {
            "parameters": [
                {
                    "$ref": "#/components/parameters/DocumentId"
                }
            ],
            "get": {
                "tags": [
                    "Contacts"
                ],
                "summary": "Get one contact",
                "operationId": "getContactById",
                "responses": {
                    "200": {
                        "description": "Contact retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Contact"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/BadRequest"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                }
            },
            "put": {
                "tags": [
                    "Contacts"
                ],
                "summary": "Update a contact",
                "operationId": "updateContact",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ContactInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Contact updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Contact"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/BadRequest"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    }
                },
                "security": [
                    {
                        "sessionCookie": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Contacts"
                ],
                "summary": "Delete a contact",
                "operationId": "deleteContact",
                "responses": {
                    "204": {
                        "description": "Contact deleted successfully"
                    },
                    "400": {
                        "$ref": "#/components/responses/BadRequest"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    }
                },
                "security": [
                    {
                        "sessionCookie": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "parameters": {
            "DocumentId": {
                "name": "id",
                "in": "path",
                "required": true,
                "description": "MongoDB document ID",
                "schema": {
                    "type": "string",
                    "pattern": "^[a-fA-F0-9]{24}$"
                }
            }
        },
        "responses": {
            "BadRequest": {
                "description": "Invalid ID or request data",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        }
                    }
                }
            },
            "NotFound": {
                "description": "Document not found",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        }
                    }
                }
            },
            "ServerError": {
                "description": "Internal server error",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        }
                    }
                }
            },
            "Unauthorized": {
                "description": "An authenticated GitHub session is required",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/AuthenticationError"
                        }
                    }
                }
            }
        },
        "schemas": {
            "Error": {
                "type": "object",
                "properties": {
                    "error": {
                        "type": "string",
                        "example": "Unable to process the request"
                    }
                }
            },
...
            "ContactInput": {
                "type": "object",
                "required": [
                    "name",
                    "email",
                    "phone",
                    "role",
                    "companyId"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Jordan Smith",
                        "minLength": 1
                    },
                    "email": {
                        "type": "string",
                        "example": "jordan@example.com",
                        "minLength": 1,
                        "format": "email"
                    },
                    "phone": {
                        "type": "string",
                        "example": "208-555-0100",
                        "minLength": 1
                    },
                    "role": {
                        "type": "string",
                        "example": "Recruiter",
                        "minLength": 1
                    },
                    "companyId": {
                        "type": "string",
                        "example": "6aab43653f32b48f16367f8d",
                        "minLength": 1,
                        "pattern": "^[a-fA-F0-9]{24}$",
                        "description": "ID of an existing company"
                    },
                    "notes": {
                        "type": "string",
                        "example": "Test contact for Acme"
                    }
                }
            },
            "Contact": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/ContactInput"
                    },
                    {
                        "type": "object",
                        "properties": {
                            "_id": {
                                "type": "string",
                                "example": "68cb1234567890abcdef1234"
                            },
                            "createdAt": {
                                "type": "string",
                                "format": "date-time"
                            },
                            "updatedAt": {
                                "type": "string",
                                "format": "date-time"
                            }
                        }
                    }
                ]
            },
...
            "AuthenticationError": {
                "type": "object",
                "properties": {
                    "error": {
                        "type": "string",
                        "example": "Authentication required"
                    },
                    "login": {
                        "type": "string",
                        "example": "/auth/github"
                    }
                }
            },
...
        },
        "securitySchemes": {
            "sessionCookie": {
                "type": "apiKey",
                "in": "cookie",
                "name": "job-tracker.sid",
                "description": "Sign in through /auth/github in this browser. The session cookie is sent automatically to same-origin requests."
            }
        }
    }
}